requirements.txt
boto3>=1.28.0
requests>=2.31.0
pandas>=2.0.0
schedule>=1.2.0
python-dotenv>=1.0.0

# Ajouter Airflow
apache-airflow==2.8.0
apache-airflow-providers-amazon==8.0.0
werkzeug==2.3.0